chore: Update babel.config.js and package.json
The changes in this commit update the babel.config.js file to add the "nativewind/babel" plugin and update the presets to use double quotes instead of single quotes. Additionally, the package.json file is updated to add the "expo-constants", "expo-linking", "expo-router", and "nativewind" dependencies, and update the "react-native" and "react-native-safe-area-context" dependencies.